+++ lib/kuby/docker/timestamped_image.rb
@@ -78,14 +78,32 @@ module Kuby
         tag
       end
 
-      T::Sig::WithoutRuntime.sig { params(build_args: T::Hash[String, String], docker_args: T::Array[String], context: T.nilable(String)).void }
-      def build(build_args = {}, docker_args = [], context: nil)
-        docker_cli.build(self, build_args: build_args, docker_args: docker_args, context: context)
+      T::Sig::WithoutRuntime.sig {
+        params(
+          build_args: T::Hash[String, String],
+          docker_args: T::Array[String],
+          context: T.nilable(String),
+          cache_from: T.nilable(String)
+        ).void
+      }
+      def build(build_args = {}, docker_args = [], context: nil, cache_from: nil)
+        docker_cli.build(
+          self,
+          build_args: build_args,
+          docker_args: docker_args,
+          context: context,
+          cache_from: cache_from
+        )
       end
 
       T::Sig::WithoutRuntime.sig { params(tag: String).void }
       def push(tag)
         docker_cli.push(image_url, tag)
+      end
+
+      T::Sig::WithoutRuntime.sig { params(tag: String).void }
+      def pull(tag)
+        docker_cli.pull(image_url, tag)
       end
 
       T::Sig::WithoutRuntime.sig { returns(T::Boolean) }